[wpilibc] Move Preferences to preferences folder
