+++ wpilibc/src/main/native/cpp/preferences/Preferences.cpp
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-#include "wpi/util/Preferences.hpp"
+#include "wpi/preferences/Preferences.hpp"
 
 #include <format>
 #include <memory>
